spotify search update error component added skip for startup_script

+++ dischord/startup_script.py
@@ -10,17 +10,28 @@ os.system("npm install")
 
 # Prompt the user for API keys
 # Save this input to the environment variables
-spotify_client_ID = input("Enter your spotify client ID: ")
-spotify_client_secret = input("Enter your spotify client secret: ")
 
-firebase_api_key = input("Enter your firebase API key: ")
 
 # combine the input with the environment variables
 # write to .env file
-with open('.env', 'w') as f:
-    f.write(f'REACT_APP_SPOTIFY_CLIENT_ID={spotify_client_ID}\n')
-    f.write(f'REACT_APP_SPOTIFY_CLIENT_SECRET={spotify_client_secret}\n')
-    f.write(f'REACT_APP_FIREBASE_API_KEY={firebase_api_key}\n')
+if not os.path.exists('.env'):
+    spotify_client_ID = input("Enter your spotify client ID: ")
+    spotify_client_secret = input("Enter your spotify client secret: ")
+    firebase_api_key = input("Enter your firebase API key: ")
+    with open('.env', 'w') as f:
+        f.write(f'REACT_APP_SPOTIFY_CLIENT_ID={spotify_client_ID}\n')
+        f.write(f'REACT_APP_SPOTIFY_CLIENT_SECRET={spotify_client_secret}\n')
+        f.write(f'REACT_APP_FIREBASE_API_KEY={firebase_api_key}\n')
+else:
+    skip_prompt = input(".env file already exists. Do you want to skip writing to .env file? (y/n): ")
+    if skip_prompt.lower() != 'y':
+        spotify_client_ID = input("Enter your spotify client ID: ")
+        spotify_client_secret = input("Enter your spotify client secret: ")
+        firebase_api_key = input("Enter your firebase API key: ")
+        with open('.env', 'w') as f:
+            f.write(f'REACT_APP_SPOTIFY_CLIENT_ID={spotify_client_ID}\n')
+            f.write(f'REACT_APP_SPOTIFY_CLIENT_SECRET={spotify_client_secret}\n')
+            f.write(f'REACT_APP_FIREBASE_API_KEY={firebase_api_key}\n')
 
 # Start the first process
 p1 = subprocess.Popen("npm start", shell=True)
